+++ src/sink/kv.rs
@@ -22,6 +22,7 @@ use halfbrown::HashMap;
 use serde::Deserialize;
 use sled::{CompareAndSwapError, IVec};
 use std::boxed::Box;
+use tremor_pipeline::EventIdGenerator;
 use tremor_value::literal;
 
 #[derive(Deserialize, Debug, Clone)]
@@ -31,6 +32,8 @@ pub struct Config {
 pub struct Kv {
     sink_url: TremorUrl,
     event_origin_uri: EventOriginUri,
+    idgen: EventIdGenerator,
+    reply_tx: Sender<Reply>,
     db: sled::Db,
 }
 
@@ -47,14 +50,11 @@ fn decode(mut v: Option<IVec>, codec: &mut dyn Codec, ingest_ns: u64) -> Result<
     }
 }
 
-fn ok(v: Value<'static>) -> Value<'static> {
-    let mut reply = Value::object_with_capacity(1);
-    if reply.insert("ok", v).is_ok() {
-        reply
-    } else {
-        // ALLOW: We know this never can happen since `reply` is created as an object
-        unreachable!()
-    }
+fn ok(k: Vec<u8>, v: Value<'static>) -> Value<'static> {
+    literal!({
+        "key": Value::Bytes(k.into()),
+        "ok": v
+    })
 }
 
 impl Kv {
@@ -65,37 +65,33 @@ impl Kv {
         ingest_ns: u64,
     ) -> Result<Value<'static>> {
         match cmd {
-            Command::Get { key } => decode(self.db.get(key)?, codec, ingest_ns).map(ok),
-            Command::Put { key, value } => {
-                decode(self.db.insert(key, codec.encode(value)?)?, codec, ingest_ns).map(ok)
+            Command::Get { key } => {
+                decode(self.db.get(&key)?, codec, ingest_ns).map(|v| ok(key, v))
             }
-            Command::Delete { key } => decode(self.db.remove(key)?, codec, ingest_ns).map(ok),
+            Command::Put { key, value } => decode(
+                self.db.insert(&key, codec.encode(value)?)?,
+                codec,
+                ingest_ns,
+            )
+            .map(|v| ok(key, v)),
+            Command::Delete { key } => {
+                decode(self.db.remove(&key)?, codec, ingest_ns).map(|v| ok(key, v))
+            }
+
             Command::Cas { key, old, new } => {
                 if let Err(CompareAndSwapError { current, proposed }) = self.db.compare_and_swap(
-                    key,
+                    &key,
                     old.map(|v| codec.encode(v)).transpose()?,
                     new.map(|v| codec.encode(v)).transpose()?,
                 )? {
-                    let mut err = Value::object_with_capacity(2);
-                    err.insert(
-                        "current",
-                        current.map(|v| {
-                            let v: &[u8] = &v;
-                            Value::Bytes(Cow::from(Vec::from(v)))
-                        }),
-                    )?;
-                    err.insert(
-                        "proposed",
-                        proposed.map(|v| {
-                            let v: &[u8] = &v;
-                            Value::Bytes(Cow::from(Vec::from(v)))
-                        }),
-                    )?;
-                    let mut reply = Value::object_with_capacity(1);
-                    reply.insert("error", err)?;
-                    Ok(reply)
+                    Ok(literal!({
+                        "key": Value::Bytes(key.into()),
+                        "error": {
+                            "current": decode(current, codec, ingest_ns)?,
+                            "proposed": decode(proposed, codec, ingest_ns)?                        }
+                    }))
                 } else {
-                    Ok(ok(Value::null()))
+                    Ok(ok(key, Value::null()))
                 }
             }
             Command::Scan { start, end } => {
@@ -107,15 +103,15 @@ impl Kv {
                 };
                 let mut res = Vec::with_capacity(32);
                 for e in i {
-                    let mut kv = Value::object_with_capacity(2);
-
                     let (key, e) = e?;
                     let key: &[u8] = &key;
-                    kv.insert("key", Value::Bytes(key.to_vec().into()))?;
-                    kv.insert("value", decode(Some(e), codec, ingest_ns)?)?;
-                    res.push(kv)
+                    let value = literal!({
+                        "key": Value::Bytes(key.to_vec().into()),
+                        "value": decode(Some(e), codec, ingest_ns)?
+                    });
+                    res.push(value)
                 }
-                Ok(Value::from(res))
+                Ok(literal!({ "ok": res }))
             }
         }
     }
@@ -134,9 +130,13 @@ impl offramp::Impl for Kv {
                 port: None,
                 path: config.dir.split('/').map(ToString::to_string).collect(),
             };
+            // dummy
+            let (dummy_tx, _) = async_channel::bounded(1);
 
             Ok(SinkManager::new_box(Kv {
                 sink_url: TremorUrl::from_onramp_id("kv")?, // dummy value
+                idgen: EventIdGenerator::new(0),
+                reply_tx: dummy_tx, // dummy, will be replaced in init
                 event_origin_uri,
                 db,
             }))
@@ -173,7 +173,7 @@ enum Command<'v> {
     /// ```json
     /// {"cas": {
     ///    "key": "key"
-    ///    "old": "<value|null|not-se>",
+    ///    "old": "<value|null|not-set>",
     ///    "new": "<value|null|not-set>",
     /// }
     /// ```
@@ -184,8 +184,31 @@ enum Command<'v> {
     },
 }
 
+impl<'v> Command<'v> {
+    fn op_name(&self) -> &'static str {
+        match self {
+            Command::Get { .. } => "get",
+            Command::Put { .. } => "put",
+            Command::Delete { .. } => "delete",
+            Command::Scan { .. } => "scan",
+            Command::Cas { .. } => "cas",
+        }
+    }
+
+    fn key(&self) -> Option<Vec<u8>> {
+        match self {
+            Command::Get { key, .. }
+            | Command::Put { key, .. }
+            | Command::Delete { key }
+            | Command::Cas { key, .. } => Some(key.clone()),
+            Command::Scan { .. } => None,
+        }
+    }
+}
+
 #[async_trait::async_trait]
 impl Sink for Kv {
+    #[allow(clippy::clippy::too_many_lines, clippy::option_if_let_else)]
     async fn on_event(
         &mut self,
         _input: &str,
@@ -196,65 +219,118 @@ impl Sink for Kv {
         let mut r = Vec::with_capacity(10);
         let ingest_ns = tremor_common::time::nanotime();
 
-        let mut cmds = event.value_meta_iter().filter_map(|(v, m)| {
-            let r = if let Some(g) = v.get("get") {
-                Some(Command::Get {
-                    key: g.get_bytes("key")?.to_vec(),
-                })
+        let cmds = event.value_meta_iter().map(|(v, m)| {
+            let command_res = if let Some(g) = v.get("get") {
+                g.get_bytes("key")
+                    .ok_or_else(|| "Missing or invalid `key` field".into())
+                    .map(|key| Command::Get { key: key.to_vec() })
             } else if let Some(p) = v.get("put") {
-                Some(Command::Put {
-                    key: p.get_bytes("key")?.to_vec(),
-                    value: p.get("value")?,
-                })
+                p.get_bytes("key")
+                    .ok_or_else(|| "Missing or invalid `key` field".into())
+                    .and_then(|key| {
+                        p.get("value")
+                            .map(|value| Command::Put {
+                                key: key.to_vec(),
+                                value,
+                            })
+                            .ok_or_else(|| "Missing `value` field".into())
+                    })
             } else if let Some(c) = v.get("cas") {
-                Some(Command::Cas {
-                    key: c.get_bytes("key")?.to_vec(),
-                    old: c.get("old"),
-                    new: c.get("new"),
-                })
+                c.get_bytes("key")
+                    .ok_or_else(|| "Missing or invalid `key` field".into())
+                    .map(|key| Command::Cas {
+                        key: key.to_vec(),
+                        old: c.get("old"),
+                        new: c.get("new"),
+                    })
             } else if let Some(d) = v.get("delete") {
-                Some(Command::Delete {
-                    key: d.get_bytes("key")?.to_vec(),
-                })
+                d.get_bytes("key")
+                    .ok_or_else(|| "Missing or invalid `key` field".into())
+                    .map(|key| Command::Delete { key: key.to_vec() })
             } else if let Some(s) = v.get("scan") {
-                Some(Command::Scan {
+                Ok(Command::Scan {
                     start: s.get_bytes("start").map(|v| v.to_vec()),
                     end: s.get_bytes("end").map(|v| v.to_vec()),
                 })
             } else {
-                error!("failed to decode command: {}", v);
-                None
+                Err(format!("Invalid KV command: {}", v))
             };
-            r.map(|r| (r, m.get("correlation")))
+            (
+                command_res.map_err(|e| ErrorKind::KvError(e).into()),
+                m.get("correlation"),
+            )
         });
-        let first = cmds.next();
-        if let Some((cmd, correlation)) = first {
-            for (cmd, correlation) in cmds {
-                let meta = correlation
-                    .map(|c| literal!({ "correlation": c.clone_static() }))
-                    .unwrap_or_default();
-
-                let data = self.execute(cmd, codec, ingest_ns)?;
-                let e = Event {
-                    data: (data, meta).into(),
-                    origin_uri: Some(self.event_origin_uri.clone()),
-                    ..Event::default()
-                };
-                r.push(Reply::Response(OUT, e))
-            }
-            let meta = correlation
-                .map(|c| literal!({ "correlation": c.clone_static() }))
-                .unwrap_or_default();
-
-            let data = self.execute(cmd, codec, ingest_ns)?;
-            let e = Event {
-                data: (data, meta).into(),
-                origin_uri: Some(self.event_origin_uri.clone()),
-                ..Event::default()
+        let mut first_error = None;
+        // note: we always try to execute all commands / handle all errors.
+        //       we might want to exit early in some cases though
+        for (cmd_or_err, correlation) in cmds {
+            let executed = match cmd_or_err {
+                Ok(cmd) => {
+                    let name = cmd.op_name();
+                    let key = cmd.key();
+                    self.execute(cmd, codec, ingest_ns)
+                        .map(|res| (name, res))
+                        .map_err(|e| (Some(name), key, e))
+                }
+                Err(e) => Err((None, None, e)),
             };
-            r.push(Reply::Response(OUT, e))
+            match executed {
+                Ok((op, data)) => {
+                    let mut id = self.idgen.next_id();
+                    id.track(&event.id);
+
+                    let mut meta = Value::object_with_capacity(2);
+                    meta.try_insert("kv", literal!({ "op": op }));
+                    if let Some(correlation) = correlation {
+                        meta.try_insert("correlation", correlation.clone_static());
+                    }
+                    let e = Event {
+                        id,
+                        ingest_ns,
+                        data: (data, meta).into(),
+                        origin_uri: Some(self.event_origin_uri.clone()),
+                        ..Event::default()
+                    };
+                    r.push(Reply::Response(OUT, e))
+                }
+                Err((op, key, e)) => {
+                    // send ERR response and log err
+                    let mut id = self.idgen.next_id();
+                    id.track(&event.id);
+                    let data = literal!({
+                        "key": key.map_or_else(Value::null, |v| Value::Bytes(v.into())),
+                        "error": e.to_string(),
+                    });
+                    let mut meta = Value::object_with_capacity(3);
+                    meta.try_insert("kv", literal!({ "op": op }));
+                    meta.try_insert("error", e.to_string());
+                    if let Some(correlation) = correlation {
+                        meta.try_insert("correlation", correlation.clone_static());
+                    }
+                    let err_event = Event {
+                        id,
+                        ingest_ns,
+                        data: (data, meta).into(),
+                        origin_uri: Some(self.event_origin_uri.clone()),
+                        ..Event::default()
+                    };
+                    r.push(Reply::Response(ERR, err_event));
+                    first_error.get_or_insert(e);
+                }
+            }
         }
-        Ok(Some(r))
+        if let Some(e) = first_error {
+            // send away all response events asynchronously before
+            for reply in r {
+                if let Err(e) = self.reply_tx.send(reply).await {
+                    error!("[Sink::{}] Error sending error reply: {}", self.sink_url, e);
+                }
+            }
+            // trigger CB fail
+            Err(e)
+        } else {
+            Ok(Some(r))
+        }
     }
 
     async fn on_signal(&mut self, _signal: Event) -> ResultVec {
@@ -270,10 +346,12 @@ impl Sink for Kv {
         _codec_map: &HashMap<String, Box<dyn Codec>>,
         _processors: Processors<'_>,
         _is_linked: bool,
-        _reply_channel: Sender<sink::Reply>,
+        reply_channel: Sender<sink::Reply>,
     ) -> Result<()> {
         self.event_origin_uri.uid = sink_uid;
         self.sink_url = sink_url.clone();
+        self.idgen.set_source(sink_uid);
+        self.reply_tx = reply_channel;
         Ok(())
     }